Move require_upload_step decorator to decorators.py, make dedicated test file
