+++ app/recordtransfer/tests/unit/views/test_media.py
@@ -1,18 +1,14 @@
-import json
 import logging
 from unittest.mock import MagicMock, patch
 
-from django.contrib.sessions.backends.db import SessionStore
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django.forms import ValidationError
-from django.http import HttpRequest, JsonResponse
-from django.test import RequestFactory, TestCase
+from django.test import TestCase
 from django.urls import reverse
 from django.utils.translation import gettext
 
 from recordtransfer.enums import TransferStep
 from recordtransfer.models import TempUploadedFile, UploadSession, User
-from recordtransfer.views.media import require_upload_step
 
 
 class TestCreateUploadSessionView(TestCase):
@@ -450,64 +446,3 @@ class TestUploadedFileView(TestCase):
         """Tear down test environment."""
         TempUploadedFile.objects.all().delete()
         UploadSession.objects.all().delete()
-
-
-class TestRequireUploadStepDecorator(TestCase):
-    """Tests for the require_upload_step decorator."""
-
-    def setUp(self) -> None:
-        """Set up the test case."""
-        self.factory = RequestFactory()
-
-    def test_request_without_wizard_data(self) -> None:
-        """Requests without wizard data should be forbidden."""
-
-        @require_upload_step
-        def dummy_view(request: HttpRequest, *args, **kwargs):
-            return JsonResponse({"success": True})
-
-        for method in ["get", "post", "put", "delete"]:
-            request = getattr(self.factory, method)("/dummy-url/")
-            request.session = SessionStore()
-            response = dummy_view(request)
-            self.assertEqual(response.status_code, 403)
-            self.assertEqual(
-                json.loads(response.content),
-                {"error": gettext("Uploads are only permitted during the file upload step")},
-            )
-
-    def test_request_with_wrong_step(self) -> None:
-        """Requests with the wrong wizard step should be forbidden."""
-
-        @require_upload_step
-        def dummy_view(request: HttpRequest, *args, **kwargs):
-            return JsonResponse({"success": True})
-
-        for method in ["get", "post", "put", "delete"]:
-            request = getattr(self.factory, method)("/dummy-url/")
-            request.session = SessionStore()
-            request.session["wizard_transfer_form_wizard"] = {"step": "some_other_step"}
-            response = dummy_view(request)
-            self.assertEqual(response.status_code, 403)
-            self.assertEqual(
-                json.loads(response.content),
-                {"error": gettext("Uploads are only permitted during the file upload step")},
-            )
-
-    def test_request_with_correct_step(self) -> None:
-        """Requests with the correct wizard step should pass through."""
-
-        @require_upload_step
-        def dummy_view(request: HttpRequest, *args, **kwargs):
-            return JsonResponse({"success": True})
-
-        for method in ["get", "post", "put", "delete"]:
-            request = getattr(self.factory, method)("/dummy-url/")
-            request.session = SessionStore()
-            request.session["wizard_transfer_form_wizard"] = {
-                "step": TransferStep.UPLOAD_FILES.value
-            }
-            request.session.save()
-            response = dummy_view(request)
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(json.loads(response.content), {"success": True})
